Testing new approach to fog lighting.

+++ shaders/lib/util/BufferFormat.glsl
@@ -4,5 +4,5 @@
   const int colortex2Format = RGB32F;
   const int colortex3Format = RGB32F;
   const int colortex4Format = RGB32F;
-  const int colortex5Format = RGBA32F;
+  const int colortex5Format = RGBA16;
 */
